Turn NokogiriDocument into Page class
MemberPage will need to know the url for the source page.

We were passing the nokogiri node and so MemberPage did not
know the URL of the source in question.

Instead, we pass the URL and Page converts it into a nokogiri
document so the URL is available to MembePage.